[Gold III] Title: 세 용액, Time: 32 ms, Memory: 2064 KB -BaekjoonHub

+++ 백준/Gold/2473. 세 용액/세 용액.cc
@@ -2,7 +2,7 @@
 using namespace std;
 using ll = long long;
 
-const ll INF = 1000000005;
+const int INF = 1000000005;
 ll n, arr[5005];
 vector<ll> ans;
